week5/backend/pyproject.toml
[project]
name = "week5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.10.0",
    "mcp[cli]>=1.6.0",
    "uvicorn>=0.34.1",
]
